requirements.txt
Django==5.2
django-ninja==1.4.1
PyJWT==2.10.1
cryptography==44.0.2
psycopg[binary]
gunicorn==23.0.0
requests==2.32.3
